cex: convert to TS
